+++ client/composables/axios/useAxiosMiddlewares.ts
@@ -1,25 +1,12 @@
 import type { AxiosMiddlewares } from './types'
-import type { Tokens } from '~/stores/useAuth'
 
-function getTokensFromLocalStorage(): Tokens | null {
-  const auth = localStorage.getItem('auth')
-  if (!auth)
-    return null
-
-  const parsed = JSON.parse(auth)
-  if (!parsed.tokens || !parsed.tokens.accessToken || !parsed.tokens.refreshToken)
-    return null
-
-  return parsed.tokens
-}
+export const skipRefreshRoutes = ['/refresh', '/login', '/users/signup', '/upload/token']
 
 export function useAxiosMiddlewares(): ComputedRef<AxiosMiddlewares> {
-  // const toast = useToast()
-
   return computed(() => ({
     requestMiddlewares: [
       {
-        route: '/refresh',
+        route: skipRefreshRoutes,
         negateRoute: true,
         onFulfilled: (config) => {
           const tokens = getTokensFromLocalStorage()
@@ -45,32 +32,6 @@ export function useAxiosMiddlewares(): ComputedRef<AxiosMiddlewares> {
       },
     ],
     responseMiddlewares: [
-      {
-        route: /^\/(refresh|login|users\/signup|upload\/token)/,
-        negateRoute: true,
-        onRejected: async (error) => {
-          if (error.response?.status === 401) {
-            const { refresh } = useAuth()
-            const refreshed = await refresh()
-            if (refreshed) {
-              const axiosInstance = useAxiosInstance()
-              error.config!.headers.Authorization = `Bearer ${getTokensFromLocalStorage()?.accessToken}`
-              return axiosInstance.value(error.config!)
-            }
-            else {
-              // toast.add({
-              //   severity: 'error',
-              //   summary: 'Session',
-              //   detail: 'Session expired, please login again',
-              //   life: 5000,
-              // })
-              navigateTo('/login')
-              return Promise.reject(error)
-            }
-          }
-          return Promise.resolve()
-        },
-      },
     ],
   }))
 }
