+++ client/composables/useAPI.ts
@@ -4,7 +4,7 @@ import type { AxiosResponse, RawAxiosRequestConfig } from 'axios'
 
 export default function useAPI<T = any, R = AxiosResponse<T>, D = any>(
   url: string,
-  config?: RawAxiosRequestConfig<D>,
+  config?: RawAxiosRequestConfig<D> & { skipAuthRefresh?: boolean },
   options?: UseAxiosOptions<T>,
 ): StrictUseAxiosReturn<T, R, D> & PromiseLike<StrictUseAxiosReturn<T, R, D>> {
   const axiosInstance = useAxiosInstance()
