+++ client/env.config.json
@@ -1,5 +1,5 @@
 {
   "public": {
-    "apiBaseUrl": "http://localhost:8080/api"
+    "apiBaseUrl": "/api"
   }
 }
