+++ client/env.dev.config.json
@@ -1,5 +1,5 @@
 {
   "public": {
-    "apiBaseUrl": "/api"
+    "apiBaseUrl": "http://localhost:8080/api"
   }
 }
