+++ client/package.json
@@ -33,6 +33,7 @@
     "@vueuse/nuxt": "^9.13.0",
     "async-validator": "^4.2.5",
     "axios": "^1.3.5",
+    "axios-auth-refresh": "^3.3.6",
     "change-case": "^4.1.2",
     "class-variance-authority": "^0.5.1",
     "eslint": "^8.38.0",
